Moved core.util to lib.util
