+++ package/CHANGELOG
@@ -47,6 +47,7 @@ The rules for this file:
   * Moved MDAnalysis.KDTree to MDAnalysis.lib.KDTree (Issue #287)
   * Moved MDAnalysis.core.transformations to MDAnalysis.lib.transformations
     (Issue #287)
+  * Moved MDAnalysis.core.util to MDAnalysis.lib.util (Issue #287)
   * base.Reader now defines __iter__
   * __iter__ removed from many Readers (now use base.Reader implementation)
   * numpy >= 1.5 required
