+++ package/MDAnalysis/coordinates/GRO.py
@@ -31,7 +31,7 @@ import numpy as np
 
 import MDAnalysis
 from . import base
-import MDAnalysis.core.util as util
+import MDAnalysis.lib.util as util
 from .core import triclinic_box, triclinic_vectors
