+++ package/MDAnalysis/lib/transformations.py
@@ -197,7 +197,7 @@ import math
 import numpy
 from numpy.linalg import norm
 
-from MDAnalysis.core.util import angle as vecangle
+from .util import angle as vecangle
 
 def identity_matrix():
     """Return 4x4 identity/unit matrix.
