+++ package/MDAnalysis/topology/XYZParser.py
@@ -34,7 +34,7 @@ Classes
 from __future__ import absolute_import
 
 from ..core.AtomGroup import Atom
-from ..core.util import openany
+from ..lib.util import openany
 from .core import get_atom_mass, guess_atom_charge, guess_atom_element
 from .base import TopologyReader
