--- package/doc/sphinx/source/documentation_pages/core/util.rst
@@ -1,2 +1,0 @@
-.. automodule:: MDAnalysis.core.util
-
