+++ package/doc/sphinx/source/documentation_pages/core_modules.rst
@@ -22,6 +22,5 @@ parsing in :mod:`MDAnalysis.core.Selection`.
    core/distances
    core/qcprot
    core/units
-   core/util
    core/log
    core/parallel
