+++ package/doc/sphinx/source/documentation_pages/lib/util.rst
@@ -1,0 +1,2 @@
+.. automodule:: MDAnalysis.lib.util
+
