+++ package/doc/sphinx/source/documentation_pages/lib_modules.rst
@@ -11,9 +11,13 @@ no error checking done on inputs so should be used with caution.
 :mod:`MDAnalysis.lib.transformations` contains a multitude of
 matrix operations for manipulating coordinate data.
 
+:mod:`MDAnalysis.lib.util` contains various file, string
+and mathematical utility functions.
+
 .. toctree::
    :maxdepth: 1
 
    ./lib/_distances
    ./lib/KDTree_modules
    ./lib/transformations
+   ./lib/util
